+++ decidim-surveys/app/commands/decidim/surveys/admin/update_survey.rb
@@ -50,7 +50,8 @@ module Decidim
           update_nested_model(form_question, question_attributes, @survey.questions) do |question|
             form_question.answer_options.each do |form_answer_option|
               answer_option_attributes = {
-                body: form_answer_option.body
+                body: form_answer_option.body,
+                free_text_option: form_answer_option.free_text_option
               }
 
               update_nested_model(form_answer_option, answer_option_attributes, question.answer_options)