+++ decidim-surveys/app/forms/decidim/surveys/admin/survey_question_answer_option_form.rb
@@ -8,6 +8,7 @@ module Decidim
         include TranslatableAttributes
 
         attribute :deleted, Boolean, default: false
+        attribute :free_text_option, Boolean
 
         translatable_attribute :body, String