+++ decidim-surveys/app/forms/decidim/surveys/survey_answer_choice_form.rb
@@ -5,6 +5,7 @@ module Decidim
     # This class holds a Form to update survey question answer options
     class SurveyAnswerChoiceForm < Decidim::Form
       attribute :body, String
+      attribute :custom_body, String
       attribute :decidim_survey_answer_option_id, Integer
 
       validates :decidim_survey_answer_option_id, presence: true