+++ decidim-surveys/db/migrate/20180405015258_add_free_text_option_to_survey_answer_options.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddFreeTextOptionToSurveyAnswerOptions < ActiveRecord::Migration[5.1]
+  def change
+    add_column :decidim_surveys_survey_answer_options, :free_text_option, :boolean
+  end
+end